expose gradual type inferface to HFtracer and generate constraints for gt nodes (#80744)
Define the function `transform_all_constraints_trace_time` which takes a node which represents a conditional, a graph and a root which represent the trace so far. We then generate constraints for the current trace, and a separate constraint for the conditional. We then combine the constraints with the node constraint in one set and we combine the constraint with the negation of the node's constraint in another set. We then transform both sets to Z3 format and we return both sets.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/80744
Approved by: https://github.com/jansel