[FSDP] Remove the dependency of ``_symbolic_trace`` in ``wrap`` (#81443)
Same as #81339, this is used to fix internal tests where ``torch.fx`` is not available in the module and ``wrap`` of FSDP is imported. With this there should be no import errors for FSDP when ``torch.fx`` is not available.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/81443
Approved by: https://github.com/awgu, https://github.com/zhaojuanmao, https://github.com/Neilblaze