Re-enable iOS simulator tests (#82027)
Issue #81613

To enable the rest of the tests, I had to disable dynamic_quant_ops test.

For further investigation and re-enabling dynamic_quant_ops (in a separate PR) need to look for a PR that went into nightly for Jul 16 2022 and broke the model file generation.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/82027
Approved by: https://github.com/huydhn