Temporary fix to skip NVIDIA driver installation from RHEL repo (#85569)
This is a temporary fix until torchrec and FBGEMM are updated to use PyTorch NVIDIA installation script instead of using the latest driver from RHEL repo.  It might take a day or so to finish updating the 2 repos, so I want to have this in place to avoid any issue with NVIDIA driver till then.  The driver from RHEL repo `515.65.01` is even newer than what we are using in PyTorch CI `515.57`.  So everything should just work with both of them
Pull Request resolved: https://github.com/pytorch/pytorch/pull/85569
Approved by: https://github.com/clee2000