[FSDP][4/N] Rework FPW test to not use FPW (#87112)
Testing coverage is pretty much preserved except that we do not test on CPU, which is not a tangible loss for FSDP anyway.

I renamed a few tests slightly, and I moved some helpers to be immediately below the corresponding test method. This makes it a bit easier to read.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/87112
Approved by: https://github.com/zhaojuanmao