[fx] Add Graph option for replace_pattern (#112409)
Summary:
Allow doing pattern replacement with just an fx.Graph instead of a fx.GraphModule,
which can let callers avoid paying the cost of `recompile()` for a small graph if they
don't need the module.

This is a significant speedup if you use hundreds of small patterns for replacement.

Test Plan: Tested in a diff stacked on top of this: {D50756722}

Reviewed By: SherlockNoMad, angelayi

Differential Revision: D50756723

@diff-train-skip-merge

Pull Request resolved: https://github.com/pytorch/pytorch/pull/112409
Approved by: https://github.com/ZainRizvi