[PyTorch] AOTI: add ArrayRefTensor (#112115)
This adds a shim for AOTI generated code to pretend a raw array works like an AtenTensorHandle. This allows parts of AOTI that generate uses of tensors to continue to be unaware of how those tensors are allocated. See the following diff/PR for usage.

Differential Revision: [D50570252](https://our.internmc.facebook.com/intern/diff/D50570252/)

Pull Request resolved: https://github.com/pytorch/pytorch/pull/112115
Approved by: https://github.com/chenyang78, https://github.com/desertfire