Trim Dynamo shards 7->3 (#117869)
We added all of the tests we wanted for now. These fit comfortably in 3
shards (the total test time previously was 0.5 hours on each shard).
Going to decrease the number of shards to 3 so that it's less unwieldy
to work with.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/117869
Approved by: https://github.com/huydhn, https://github.com/malfet