[DCP][BC] Remove the dependency on _shard.TensorProperties (#116248)
ShardedTensor is in the maintence mode and is going to be deprecated. DCP's metadata should not rely on any definitions in ShardedTensor. This PR creates a replica of TensorProperties in DCP and removes the dependency on _shard.TensorProperties

Differential Revision: [D52357732](https://our.internmc.facebook.com/intern/diff/D52357732/)

Pull Request resolved: https://github.com/pytorch/pytorch/pull/116248
Approved by: https://github.com/wconstab, https://github.com/LucasLLC, https://github.com/wz337