[Profiler] iterate frontend function events for profiler post processing (#124596)
The `function_events` in `_parse_kineto_results` is used to contain all function events from the result. It contains 2 kinds of events. One is frontend function events whose correlation id is 0, for example, `aten::add`, `aten::mul`. They are on the top level of the profile results. The other is the backend events, which are associated with the frontend events and its correlation id is > 0, for example, `at::native::vectorized_elementwise_kernel`, it should be the backend event of a frontend element-wise op. They have the device execution duration for the related frontend op.

In the following post processing code, the **frontend function events** should be iterated to find its correlated backend events in `device_corr_map`, instead of iterating all function events, because `device_corr_map` is designed as a dict, whose key is the id of the frontend function event.
https://github.com/pytorch/pytorch/blob/c3163e5180017f8270d99332ff3734c078369a20/torch/autograd/profiler.py#L543-L560

https://github.com/pytorch/pytorch/blob/c3163e5180017f8270d99332ff3734c078369a20/torch/autograd/profiler.py#L537-L540
Pull Request resolved: https://github.com/pytorch/pytorch/pull/124596
Approved by: https://github.com/aaronenyeshi